chore: move test to tests folder
