+++ src/common/grpc/Cargo.toml
@@ -27,7 +27,6 @@ tower = "0.4"
 [dev-dependencies]
 criterion = "0.4"
 rand = "0.8"
-tempdir = "0.3"
 
 [[bench]]
 name = "bench_main"
